remove dep on ConnHelpers
This was the last plug that used the ConnHelpers, cleaning up.
