+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Unreleased
 
 - Refactor: removed AdminAuthenticator's use of ConnHelpers
+- Refactor: removed the Router's use of ConnHelpers
 - Fix: Don't persist Shop on ShopServer init
 - Admin Authenication Plug now redirects to install if shop is not found
 - New ShopServer.get_by_client_id/1
